Changes in API Service POST and GET methods due to User Steps migration
