+++ tests/e2e_tests/component/iot_api/services/list/terrain.py
@@ -1,14 +1,15 @@
 from lettuce import before, after, world
-from common.user_steps import UserSteps
+from common.functions import Functions
 
 
-user_steps = UserSteps()
+functions = Functions()
 
 
 @before.all
 def setup():
     world.remember = {}
-    
+   
+
 @after.all
 def clean(total):
-    user_steps.clean(world.remember)
+    functions.clean(world.remember)
